Move main validation logic to `validation` module
The main validation logic isn't tied to any decorators, so `decorators`
wasn't the most suitable module for it.
